Provide non-exported callers for generic iterables
When you have a vector-of-arrays, splatting introduces a tuple
of unknown length, and even when each item is of the same type
just the unknown length introduces pretty serious inferrability
problems. This provides a non-exported variant of `paddedviews`
and `sym_paddedviews` that is designed to work on containers
rather than via splatting.